feature: establishes the four-layer submodule directory
